with-tracing/requirements.txt
Flask==2.3.3
prometheus_client==0.19.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0 # For gRPC export to Tempo
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
gunicorn==22.0.0